Prologue review
Interesting. I preferred the parts after Donaar arrived. The first part is a bit fast-paced. I think it could be improved by slowing it down a little with "painting the scenery". For example, what does Sora, Ildrex and the father, Nadraa, look like? We get the location as South east of Ashkeeper Peaks and that we are in a city, but nothing to tell us whether we're in an inner city quarter or in the outskirts, etc etc. Something to place us between the rolling grass hills of the Shire and the marble and stone of Minas Tirith.